added typescript, linted pinia modules
