+++ src/stores/grade-module.ts
@@ -1,24 +1,23 @@
-import { defineStore } from 'pinia'
+import { defineStore } from 'pinia';
 
 interface State {
   grade: string
 }
 
-export const useAuthUserStore = defineStore('schedules', {
+export default defineStore('grades', {
   // convert to a function
   state: (): State => ({
     grade: 'None',
   }),
-
   actions: {
     initialize(): void {
-    if (localStorage.grade) {
+      if (localStorage.grade) {
         this.setGrade(localStorage.grade);
-        }
+      }
     },
-    setGrade(grade: string): void{
-        this.grade = grade;
-        localStorage.grade = grade;
-      },
-  }
-})
+    setGrade(grade: string): void {
+      this.grade = grade;
+      localStorage.grade = grade;
+    },
+  },
+});
